Intro else in try-except
else would condense the program by being more reflective of our actual
purpose. In this case, we only wanted to try whether the user
inputted an integer. We didn't want to also try the print function.

However, placing the print function on its own outside of the
try-except statements would always run the print function, which
would lead to a NameError since x wouldn't be assigned a value if
the user doesn't enter an int and we run into a ValueError, going
into the except statement instead of the try statement.

else circumvents this problem by acting mutually exclusive to the
except statement and only printing when the try statement fully
succeeds.