Demonstrate if __name__ == "__main__":
Try running say.py with and without "if __name__ == "__main__":"

The difference is that main() from sayings.py will be ran in say.py
if main() is run alone.

However, if main() is run with "if __name__ == "__main__":", then
main() won't be run in say.py.

The reason for this is because "if __name__ == "__main__":" ensures
that main() is only called conditionally to make sure that main()
is only called when the file is run directly (rather than run
indirectly as in the case with say.py importing the sayings.py
library). It does this by checking if the __name__ of the file is
itself "__main__", then it would actually run main(). This is because
when you import the library, __name__ is not "__main__".
- This is demonstrated by printing __name__ in sayings.py and comparing
by running sayings.py and say.py.