Demonstrate Significance of return in Unit Testing
And in general.

In the hello() function, print() is a side effect, and doesn't
actually return anything. Therefore, test_hello() would never
actually pass the test even if hello() runs "as intended". This is
because test_hello() also runs as intended and is looking for a
return value when there isn't one.

This leads to a fundamental change in the way we implement our
functions so that they have a return value, allowing our functions to
always be testable. This is best practice for larger and larger
programs. It is important to not have side effects, especially if
we want our code/functions to be testable.