Change to Automate and Test Multiple Values with a Loop
However, ideally, our tests should be very simple.

We don't want to write so much code that our tests might be flawed
and end up needing to write tests for our tests.

The goal is to keep our tests so simple that we can just eyeball
the test and understand why our tests would result in a fail/pass.