Demonstrate Running a Folder/Package of Tests
__init__.py is used to indicate to Python to treat this test folder
not just as a module but rather as a package. A package is multiple
Python modules organized into a folder.

This allows us to just do:
"pytest test"
This will allow pytest to search through the entire test folder and
run all the tests it finds in this test "package". Thus, allowing us
to run a folder of test files instead of just one test file.