Sort Dicts using a Key Function
The sorted() function allows us to sort dictionaries by passing
in as a parameter a function that returns the value of a key in a
dictionary. This value then acts as the key by which the sorted()
function will sort by. This is useful when sorting a dictionary that
has multiple keys.