Use csv Module to Better Read from CSV Files
Our previous implementation with split(",") would have raised a
ValueError if there was a comma in one of the columns/fields just
because we were splitting by commas.

E.g. Harry,Number Four, Privet Drive
Harry and Number Four, Privet Drive would be the name and home
respectively, but our previous implementation would have interpreted it
as 3 different things to split, but we were only assigning two variables
to it, thus raising an unintended ValueError -- indicating too many
values were being passed.

By using the csv module, we don't have to reinvent the wheel and come
up with our own way to circumvent this problem, but instead we can just
use what was created by others who have had this problem before.

In this code, each row in reader would split the csv rows as intended
into two variables, and a list would be returned, and thus we can access
the individual values by indexing to them with row[0] and row[1].

This standardized approach with the csv module uses quotes to clearly
identify separate comma-separated values.