Add a Header Row and use DictReader to Read from CSV
Makes our program more robust against changes. Previously, was assuming
the first column was always name and the second column home. However, if
someone were to open the CSV file in e.g. Excel (or manually), and
switch the columns, then the columns would now be reversed but our
program still assumes the first column is name and second column home.
It would be incorrect.

By using DictReader with a header row in the CSV file, we make our
program more robust by accounting for this possiblity as long as the
header row changes with the rest of the rows. We are assigning variables
based on the values in our header row rather than assuming the first
column was name and second column home. We are instead checking which
column is name and which column is home based on the header row.