Return Multiple Values and Variable Assignments with a Tuple
A tuple is immutable.
A list is mutable.
In a list, we can change the values in the list -- e.g. by indexing to
0, 1, etc. and changing those values. It is mutable.
In a tuple, we can't change the values because it is immutable.

If you don't need to change a variable with multiple values, why allow
it to be changed and invite room for error?