Intro 'def __init__(self)' Method and Objects as Instances
An object is just an instance of a class.

In classes, these are not just attributes/instance variables but also
functions/methods that you define that behaves in a special way. These
functions allow you to determine behavior in a standard way -- special
methods.

__abc__ = "Dunder"

def __init__(self):
- This dunder init method is specifically known as an instance method.
If you want to initialize the contents of an object from a class, you
define this method. Similar to adding keys to dictionaries, we're now
adding instance variables to objects.

Student(name, house)
- Here, we're treating the Student class as a function and passing in
two values -- name and house.
- This is known as a constructor call. This line of code will construct
and instantiate a Student object. It's going to use the Student class as
a template/mold so that every student is structured the same way --
having a name and a house. But because we can pass in arguments to the
Student() function, we're going to be able to specifically customize
the contents of the Student object.
- Compare: a suburban neighborhood -- all houses look very similar,
having the same blueprint, but there are specific differences between
them.
- By calling Student(name, house) as a function, suggests that there is
a function somewhere that has been defined with def and will be called.
Where is it?

In the Student class itself with 'def __init__(self, name, house):'.
- The authors of Python have defined this line as the way to implement
the initialization of an object in Python.
- self acts as the way by which we can assume attributes to the object.
Without self, we wouldn't have a SPECIFIC way to store attributes.
Literally, where would we put/store name and house in the current
object? How do we store these variables in the current object that has
just been instantiated? The solution was to create this self argument
(can technically be named anything -- 'self' is just the main
convention) that has to come first to give you access to the current
object that was just created.

When you call Student(name, house), as a constructor, it constructs
a Student object, but there's nothing in that object initially -- no
name, no house. But the object exists in the computer's memory. It's now
up to us to store that name and house inside of that object. How? Python
will now automatically call the __init__ method and pass in a reference
to an argument that represents the current object itself that it just
constructed in memory for you -- this argument is self. Now, it is up to
us to populate it with values. This means, inside the __init__ method,
we create new attributes/instance variables inside of this otherwise
empty object.
E.g. self.name = name.
- Here, we call self, which is an argument that represents the current
object itself that was just constructed, create a new attribute/instance
variable for it called name via self.name, and assign to it the value
of name, another argument that we just passed into this object method
when we constructed the Student object via Student(name, house).
- The name in self.name could have been anything, but it's much clearer
to have a properly named variable and even one that matches with the
argument that it's taking in -- name in this case. This is the
convention to store the argument's value in an identically named
instance variable.
- We've now instantiated this Student object with the name and house
instance variables.

An object is just an instance of a class.
- From https://en.wikipedia.org/wiki/Instance_(computer_science):
"In practice, this instance usually has a data structure in common with
other instances, but the values stored in the instances are separate.
Changing the values in one instance will then not interfere with the
values of some other instance. "
- An instance has a data structure in common with other instances
in this case with Python because they are instances of the same class.
- '"instance" emphasizes the distinct identity of the object.
The creation of an instance is called instantiation.'
- Important to note that Wikipedia defines what we've done here as
class-based programming, as a style of object-oriented programming.
Specifically, objects are created from classes via constructors.
Instantiation is AKA construction.

[From CS50P Lec8 Q&A]
Python calls this __init__ method the initialization method. It
initializes the value. When we call Student(name, house), we are
technically constructing the method. There is another method called
__new__ that actually handles the process of creating an empty object
in memory for us, but generally as programmers, we don't need to
manipulate the __new__ method (unlike the __init__ method which we
need to manipulate). It just works for us. Instead, we define our own
__init__ method inside of our class and that __init__ method
initializes the contents of the object. So, technically, there is a
distinction between CONSTRUCTING the object with __new__ and
INITIALIZING with __init__, but in the world of Python, we pretty much
only worry about the __init__ method. Python generally does the
constructing part for us.

Can store more than one name/house (i.e. multiple values for the same
instance variable) by storing a list, tuple, dictionary, etc. instead.
- Just apply what we already know.

Why use classes vs. dictionaries?
- You can do much more with classes.
- You can ensure the correctness of your data more with classes.
- You can error check things.
- You can generally design more complicated software more effectively.