Validate within Objects and Raise Errors
What if the user does not give a name and just hits enter when prompted
a name?
- We don't want to put in our computer's memory a bogus Student object
that no name. Ideally, we would want to check for errors before we even
create the object, so we don't create a nameless student. It would be a
bug to have an object with no name

Similarly, we don't want the user to type in something random for their
house (at least in the world of Harry Potter). There's only 4 valid
houses at Hogwarts. It would be nice to validate that the user's input
is at least in that list.

We could do that validation in the get_student() function
- Check if the name is empty? If so, don't create Student object
- Is the house one of the 4 valid houses? If not, don't create Student
object
But doing this would be decoupled from the student itself
- get_student() currently exists as our own function in the student.py
file.

**Classes and, more generally, object-oriented programming encourages
you to encapsulate inside of a class all functionality related to that
class.**
- If you want to validate that a name exists, or validate that a house
is correct, these belong fundamentally in the class called Student
itself. Not in some random function written elsewhere.
- This is just methodology. If you think about writing longer and
longer code that gets more and more complicated, it should make
intuitive sense that if you keep all of the student related code (name
and house) in the Student class. It's just better organization -- Keep
all of the related code together.

When Student() is called, the Student object/class has already been
created.

Use raise to raise an exception/error yourself when something
exceptional (bad) happens to later catch it. Can also include a message
with the error.
E.g. raise ValueError("Missing name")
Can now combine this with try and except statements to catch the errors