Validate for House; Classes vs. Dictionaries
If you add a key-value pair to a dictionary, it's going in no matter
what -- even if the name is empty or the house is a random string of
text. However, with a class and with this __init__ method, we can now
control exactly what will be "installed" inside of this object. We have
more control over the correctness.

By raising errors, we can now catch them with try-except. We now have
more power with exceptions by catching and raising them ourselves.

You can put classes in another Python library (module/package) to later
import so that the class can be used across multiple Python projects.
You can have a library of classes. In fact, we've already been using
classes from third-party libraries.

__init__ functions
- Just like other functions generally, but special in the sense that
they are going to be called automatically by Python when creating the
class/object
- You can accept optional variables by giving the parameter variable a
default value
E.g. def __init__(self, name, house=None):

Errors
- Can create your own errors
E.g. raise ErikError("Missing name")
- Prof David didn't touch on this, but this actually doesn't work as
is. Maybe have to define the error as a new function first.