Intro __str__(self)
Without __str__, by default, print(<object>) (e.g. print(student))
prints something like <__main__.Student object at 0x102733e80>. It is
printing the location of the object in your computer's memory.

With __str__(self) defined in the class, this will define what string
to return when a function asks for a string -- e.g. a function like
print().

__str__(self)
- Only takes in one argument -- self
- Return a string

How it works
- __str__ takes in a self argument by default. It will always be passed
a reference to the current student object
- When print(student) is called, print() is hoping to get a string from
student, and so this will trigger __str__() method, and Python will
automatically pass into that method a reference to the object that's
trying to be printed so that the programmer can return a (formatted)
string
- Now, the string is customized to the specific object we're trying to
print

__str__ vs. __repr__
- str is for the users of the program; more user-friendly
- repr is for the developer; includes more information such as the type
of object, etc.