Intro Defining Own Methods in Objects/Classes
Here, we've created our own charm() method for the Student class that
we can call with our student object. We've done this by taking in
another parameter in the __init__ method called patronus, which would
then be stored in the object itself when a value is passed during the
construction of the object. This instance variable can then also be
used when we call our charm() function/method, which then performs its
task by return an emoji.

A method has to take in at least one argument that is self. This is by
convention but it also allows the method to call on itself and access
the instance variables and other methods that the object itself holds.
Self is a reference to the current method itself.

A method is called by <object>.<method>(). This is similar to accessing
instance variables from an object.
E.g. student.charm()

Difference between previous __init/str__ methods and new charm() method
- __init/str__ methods are automatically called when you call on the
object as it is. E.g. print(student) for __str__. For init and str
specifically, they are also methods already exist in Python for classes
that we now defining/specifying
- charm() method is a custom built method that we defined ourselves.
It's the same as defining a new function, just inside a class. Calling
the method is optional, not automatic, and thus has to be explicit.
E.g. student.charm().
We've used methods before as well.
E.g. <string>.lower()