Intro Properties and Decorators, Getter and Setter
With classes and objects, we can still access instance variables via
the dot (.) notation, overriding our error checking and validations.
E.g. 	student = get_student()
	student.house = "Number Four, Privet Drive"
- The get_student() validated if the house being passed when
constructing the Student object is one of the four valid houses. So
that was all done. But then, we can later override that validation and
directly change the value of house in the student object into an
invalid value by accessing the instance variable directly via
student.house. This circumvents all our validations. The validations
in the __init__ method are only called when we first create the student
object. There's nothing stopping us from changing name or house after.

Properties are an attribute that have even more defense mechanisms put
into place; a little more functionality implemented by you (the
programmer) to prevent programmers like you and others from messing
things up. A property is just an attribute that we have more control
over by writing more code to it.

@property
- property is a function in Python
- In this case, we have a decorator defining a property

Decorators
- Allows you to "decorate" functions
- Denoted by @ syntax
- Decorators are functions that modify the behaviors of other
functions

Getter
- A function/method in a class that gets some attribute
- Always accepts only one value -- self
- Logic: if we're just getting the value, you don't need to pass
anything else in because you already know the object. So, you're just
going to access the object itself and get the property in the object

Setter
- A function/method in a class that sets some value
- Always accepts two values -- self, and something else (the new value
we're setting)
- Logic: if you're trying to set the property to something else, you've
got to pass in the argument which contains the value to which you want
to set it

The purpose is to try to prevent programmers from circumventing our
error checking. To do this, we'll require that in order to ACCESS (GET)
an attribute, you have to go through some function. Let's also require
that in order to SET some attribute, you have to also go through some
function. Conventionally, these functions are called a getter function
and setter function, respectively.

Why are we using functions/methods? Once you have functions, these are
just actions/verbs that we can create ourselves. We can put any error
correction we want inside of these functions.

So, how can we prevent the user from setting house to an invalid value?
We put the error checking in the setter function.
E.g.
def house(self, house):
    if house not in [...]:
      raise ValueError("Invalid house")
    self._house = house

If we were to do this instead:
def house(self, house):
    self.house = house
- This would be the same as just immediately setting whatever the
programmer tries to set as the value for house, without any error
checks

With proper error checking with a setter function, whenever the
programmer tries to set a value to the instance variable
(e.g.student.house = "xyz"), Python will not just let the programmer
access the instance variable directly. Instead, the setter function
will be called, which will error check the value.

This happens because when the program sees student.house, where house
is also the name of the getter or setter, and then it sees the equal (=)
 sign, indicating assignment, that's enough of a cue for the program to
 prevent the programmer from access the attribute directly and instead
 go through the setter function.

@property
- A decorator that tells Python to treat the method below as a getter
- For the function itself, you name the function exactly like you would
like the property to be called -- 'house' in this case

@house.setter
- Now, a new decorator is automatically created called @house
- .setter additionally decorates the method as a setter function
- This line is a clue to Python that there will be a function below to
be used as a setter function
- The function/method has an identical name to the getter function, but
this one takes 2 arguments -- self and house. house is the (new) string
that comes from the user's input that will be checked and set if valid
- Now, we don't need to have the error checking in the __init__ method
either. Everything can be done from this one setter function -- whether
we're creating the object for the first time or assigning a new value
to the object after. The setter function will be called any time we try
to access ".house".

self._house
- Last line in the setter function so that if it passes the error
check, house would be assigned to self._house. I.e. when assigning a
value to student.house, it goes through the setter function, goes
through the error check, and finally the function will assign the value
of house to the instance variable _house for the programmer (indirectly)
- The underscore (_) is used before house because we would have
collided names if we were to leave it it as 'house'.  If we have an
instance variable called name and house, we cannot also have
functions/methods called house. They're going to collide. We have to
decide -- do we want the variable to be called house, or do we want the
function to be called house?
- By convention, the fix is to add an underscore to the instance
variables to differentiate them from the methods. So now, the instance
variables are called _house, and the property (an attribute/method) is
called house.

Properties give us the ability to define getters and setters using
@ syntax. Ultimately, they're allowing us to decorate some function we
wrote with this special behavior.

Q&A
Why use getters/setters?
- Gives us even more control over our validation. We want validation at
the start but also later on, so that the instance variables cannot be
so easily, directly accessed and changed at will by the/another
programmer (by mistake or on purpose)
- It enables Python to detect when we're manually trying to set (or get)
a value

Why doesn't self.house in the __init__ method have an underscore as
well (i.e. why isn't it self._house too?)
- By having it as self.house, it becomes the same pattern that we want
our Python program to recognize so that it automatically call the
setter even when we're passing in the house via the __init __ method
(i.e. when creating the object)
- If it were self._house instead, that would circumvent the setter, and
now there's no error checking in the __init__ function. This means that
there is no error checking when constructing a new Student object