Add Property for name; Instance Variables Can Still be Changed
The instance variable can still be changed if we directly access
_house instead of just house. This is because this is now what our
instance variable is called. So all we've really done is make it
slightly harder for the programmer to directly access the instance
variable and change it on purpose/by accident in an incorrect way.
However, ultimately, you can still access it. It isn't baked into the
language itself, like Java, that just prevents you from doing things
like this. In other languages there are notions of public code that can
be accessible by anyone or private and protected where no one should be
able to change the values of these instance variables.

In Python, it's just the honor system. If a variable has an underscore
before it, don't touch it. The more underscores there are, the more you
shouldn't touch the variable (e.g. 2 underscores). In Python, there's
no public/private notions of visibility for instance variables.
Technically, there's nothing stopping us from circumventing all of
these things we've put in place (the setters and getters) to prevent
incorrect manipulation.

Q&A
Terminology:
Attribute: anything that comes after a dot (.)
- This could be an instance variable, a function (like a method or a
property)

Property
- Always a function in Python, defined by getter or setter

Decorator
- Syntactic sugar for what would otherwise be an explicit function call
- Google "Python property function"
- There IS an alternative syntax, but no one does it that way because
a decorator is meant to make things prettier and more readable

A method is only a property if you, the programmer, define it as a
property via the "@property" decorator/keyword

A decorator is just a function that gets called on another function.
The @ sign is a special syntax to just make this easy. There are other
ways to invoke decorators. The "@property" decorator comes with Python.
It just works with the language. But, you can create any decorator you
want as well. E.g. decorating routes in web applications (e.g. Flask).
Another example  of decorators - CS50's check50 program uses decorators
to indicate that a function is a 'check'.

Working with objects (storing them, reading them)
- Can store all of this data in a CSV file
- Can read from the CSV and, instead of reading the data into a
DictReader, you can read it into a list of objects - each of which is
a Student class. This is just another way of allowing us to represent
data.
- Can also store the data in a database too (e.g. web programming). We
can manipulate the data with Python code and classes, but you can also
save the data in a database.
- What we're doing today with this lesson on OOP is having more tools
to manipulate data while it's in the computer's memory (i.e. not being
stored in a file)