Demonstrate Familiar Data Types as Classes/Objects
Notice how these classes are all in lower case, whereas when we create
our own classes, we start with a capital. This is intentional.

For default Python classes, such as data types, the authors decided
to have them all in lower case -- because they are default classes.
However, the convention when creating your own class is to have it
start with a capital to differentiate that it is a new/custom class.