Intro Class Methods and Class Variables
Aside from instance variables and instant methods associated with
classes/objects, there are other types of variables and methods too.

Class methods
- Sometimes, it's not necessary/sensible to associate a function with
objects of a class but rather with the class itself.
- An instance or an object of a class is a very specific incarnation
- Recall analogy comparison classes and objects with identical houses
in a suburban neighborhood that are all a little different because of
different paint, etc. Sometimes, you may have functionality related to
each of those different houses that isn't distinct or unique for any of
the houses. It's functionality that's going to be exactly the same no
matter the house.
- Same in the world of OOP. Sometimes, you want some
functionality/action to be associated with the class itself, no matter
what the specific object's own values or instance variables are.

@classmethod
- This is a decorator (function) that can be used to specify that this
method is not by default implicitly an instant method that has access
to self (the object itself), but rather this is a class method that is
not going to have access to self but it does know what class it's
inside.

When should you use a class to represent something in your code?
- When you want to represent a real (or fantasy) world entity
- E.g. a student, a sorting hat, etc.
- Not always the case that classes represent real world entities, but
it's common as a way to keep things organized

A class is a mold or blueprint

Sometimes it suffices to know what the class is and there might not
even be objects of that class
- This happens when we don't want to create multiple objects for a
class (e.g. sorting hat in this case; there's only one sorting hat in
the Harry Potter world)
- In that sense, we can use a class as a container for data and/or
functionality that is just somehow conceptually related
- @classmethod allows us to do just this

Class variables
- Exist within the class itself. Simply a variable in the class itself,
without self.<variable> = <value>, but rather just <variable> = <value>
as we've done so many times before
- There's just one copy of the variable for all of the objects thereof.
They all "share" the same variable. This is because it's a class
variable that is inherent to the class not the objects instantiated.
This is why we don't use self anymore since self refers to the object
that would be instantiated

To define a class method
- Add the @classmethod decorator above the method/function
- No longer pass in 'self', which was a reference to the object created,
but rather, pass in 'cls', which is a reference to the class itself
- Written as 'cls' because 'class' would conflict with declaring an
actual class, so 'cls' is used instead

When accessing a class variable (or method), use cls.<variable>
- Instead of self.<variable> -- which would be how you access an
instance variable in an object
- It's now a class variable, so you access from the class itself via cls

To use a class method, no longer need to instantiate an object from the
class
- Just use the class method from the class directly, as seen in the code
E.g.
A class method:
Hat.sort("Harry")
Vs.
An instant method:
hat = Hat()
hat.sort("Harry")
- For an instant method, have to instantiate a hat object first

There is no decorator for an instant method such as @instantmethod,
unlike class methods. A method is automatically an instant method when
you define it

[Q&A]
Can you define one class inside another?
- Yes
- Generally this isn't done, but there are cases where this can be
helpful, such as for larger, more sophisticated programs

Why use classes/objects?
- For larger and more complicated programs
- We want to focus on individual ideas
- OOP gives us a way to encapsulate related data (variables) and
functionality (methods) inside of things that have names. These things
are called classes
- Libraries are another solution to a similar problem -- defining
packages/modules and put some of your data and functionality in there,
and that's fine too
- Sometimes, which one you should use overlaps. The overlapping region
means you could use either one -- a class or a package. Overtime, you
develop an instinct or preference for which tool to use. In the end,
these are all tools. There are conventions, but they're all just tools
in the end