Update student.py Using a Class Method
Inside of Student class, we now have a method called get()
- This means this method can now be called without instantiating a
Student object first. It would be weird if you had to create a Student
object in order to call get()...in order to create another Student
object. Let's just get a Student object via a class method that does
not require us to create a Student object first.

With this, we've now moved the same code from before in the
get_student() function into a class method inside the Student class in
a more simply named function called get(). Everything that's related to
Student is now encapsulated in the Student class.

We can now instantiate (create) a Student object by just using cls() --
the cls that's included as a parameter for the class method
- We could use Student() directly (e.g. return Student(name, house) in
the Student class and get() class method itself), but using cls() is
more conventional and it will both solve and avoid problems down the
line with more complicated code.
- cls() just means to create an object of the current class. What class
is that? It's whatever cls is. And that, by definition, is Student
since that's the class that cls is in. If you call Student instead,
this leaves room for human error in case you mix up classes or names.

Student.get()
- This is nicer to read than before; it tells you what's going on.
Don't need a separate function called get_student().
- The get() functionality is built into the Student class and all the
related code is built together. We know that get() is associated with
the Student class
- It's a simpler name too

[Q&A]
Does the class have to be defined before the main() function in terms
of the order of the program?
- It doesn't actually matter, as long as you call main() after
everything has been defined.
- I.e. it's ok to define main() first, then define the Student class,
then call main()
- Generally though, classes are defined at the top of the file. It
would be even cleaner to move the class definition to its own file and
then import it -- to make reusable code for other programs and others
in general by putting it into its own module/package