Intro to Inheritance & the Incentive to Use Inheritance
Through OOP, we can design our classes in a hierarchical fashion where
we can have one class inherit from or borrow attributes (methods,
variables) from another class if they all have those in common.

In the file wizard.py, we can see how name is present in both the
Student and Professor class. If we were to scale this up and add more
attributes, we could imagine how there could be even more duplicates
between the two classes such as one's patronus. There should be an
opportunity to improve this.

How about if we recognize that both Students and Professors are both
wizards/witches, and so they both have some common characteristics or
attributes with respect to being a wizard/witch.