Implement Inheritance
The Wizard class now holds the common variable that could be
conceptually inherited.

In Python, there is a special keyword that we can use inside any
sub (descendant) class. The parent class is Wizard, and the child/sub
classes are Student and Professor. The keyword is super. We refer to
the super class, meaning the one conceptually above it, in the Student
and Professor classes to refer to the Wizard class.

class Student(Wizard):
    def __init__(self, name, house):
        super().__init__(name)
       	self.house = house
In the Student (or Professor) class' __init__ method, call the super
class' (Wizard) __init__ method, passing in only name and let the super
class do it's thing, then initialize the Student object in memory with
all the name-related information. Then, in the next line, additionally
customize the Student object with the stuff that is Student specific
with self.house = house.

So with OOP and classes specifically, we have the ability to not only
encapsulate and represent real-world entities with these classes, we
can also impose hierarchy. So, if we have two classes in the world (two
real-world entities) that somehow share common functionality, we have a
way to have one inherit from the other via sub classes and super
classes.

The ... are placeholders for additional functionality in the classes
such as printing out text.

The only new idea here is: when defining a class, by adding in
parenthesis the name of a class, we can say that this new class that we
are defining should inherit all of the functionality from the super
class. And we're additionally saying with super(), that we still have
the ability to call that super class' own __init__ method to do any of
the stuff that's common to (Students and Professors).

We've actually already been using inheritance, whether we were aware of
it or not with exceptions. Exceptions are actually set up in a
hierarchical structure because the authors of Python found that there
were commonalities between exceptions and didn't want to code multiple
different exceptions with the exact same code between them. So,
exceptions were set up in a hierarchy via inheritance instead.

Inheritance is a key feature to a lot of OOP languages.

Q & A
Can we have a super-super() class work, where your hierarchy is even
taller than the 2-levels of hierarchy that we currently have?
- Yes. What's nice about inheritance is that, as the name implies, you
would inherit properties from higher in the hierarchy.
- When you descend from another class -- when you subclass a super
class or a super-super class, you inherit all of the functionality
-- not just from one level above you but from all the levels above you.
- So indeed, you can access some of that functionality as well and can
even override it if you want some of these classes to behave a little
bit differently than others.

Can you have two parents on the same level? (i.e. two super classes?)
- Yes. There are ways to implement descendants from multiple parents,
and there's different ways to do this.
- Just not taught in this course.