Intro to Operator Overloading & Incentive to Use
We can take common symbols like +, -, or other such syntax on the
keyboard, and we can implement our own interpretation. E.g. + (plus)
does not have to equal addition and - (minus) does not have to equal
subtraction. In fact, we've already seen another context in which plus
means something else. In Python, plus has not always meant addition,
per se. Plus has also meant concatenation (for joining two strings),
and for adding to a list. Plus has been overloaded by the authors of
Python for us, so we can use the same symbol (similar to the example
of plus as addition, concatenation, adding to a list) but with
different datatypes to solve slightly different problems.

What we have right now works, where we're accessing the individual
variables inside each object and then adding them into a new variable,
and then creating a new Vault object to store the total. However,
wouldn't it be nicer to be able to overload (i.e. modify/add another
functionality to) the + (plus) operator that would just allow us to
just add two Vaults together instead of accessing everything
individually and then creating a third Vault?