Introduce Scope of Local Variables
Still have an UnboundLocalError, but for a different reason.

This time balance is a local variable.
A local variable is one that exists in the context of a function.

If you declare a variable in a function, such as main() in this case,
that variable is local to that function. Other functions, such as
deposit() and withdraw(), do not have access to that same function --
because it is only local to main().

Maybe the solution is then to move the balance variable locally, but
then we can't change it as we've seen already.