Introduce global Keyword
To change the value of global variables locally in functions, we need
to use the global keyword as seen here. By using the global keyword,
we tell Python that this variable is not a local variable but a global
variable, and we want to change this global variable

[Q&A]
What happens if you declare a variable globally AND identically locally
inside a function (same name)?
- The local variable will be prioritized over the global variable.
You'll be able to use the local variable and it'll have no effect.
Temporarily, Python will only know the local variable exists.
- The general rule of thumb is to just not do that. Not only might
this create bugs in your program as you don't quite change what you
might have intended to change, but it also won't be as obvious to other
readers of the code as well.

What happens if you pass in the balance variable instead?
- If you pass in the balance variable and then change it in that
function, it'll only be changing the local copy. It's not going to be
changing what was outside of the function (from where it was passed in)
- We need a better way instead.